+++ src/defaults.js
@@ -16,20 +16,16 @@ export class NPupDefaults {
 		this.firstAnimationDuration = firstAnimationDuration;
 		this.lastAnimationDuration = lastAnimationDuration;
 	}
-
 	static getNewElementId() {
 		this.elementId += 1;
 		return this.npupElementId + this.elementId;
 	}
-
 	static setNPupBar(__NPupBar) {
 		this._NPupBar = __NPupBar;
 	}
-	
 	static addMessage(newMessage) {
 		this.messages.push(newMessage);
 	}
-	
 	static removeMessage(messageId) {
 		let deletedIndex = -1;
 		for (let i = 0; i < this.messages.length; ++i) {
@@ -44,7 +40,9 @@ export class NPupDefaults {
 		}
 		return false;
 	}
-	
+	static clearMessages() {
+		this.messages = [];
+	}
 	static isMessageExist(messageId) {
 		let deletedIndex = -1;
 		for (let i = 0; i < this.messages.length; ++i) {
@@ -58,7 +56,6 @@ export class NPupDefaults {
 		}
 		return false;
 	}
-
 	static close(messageId) {
 		if (this.removeMessage(messageId)) {
 			this._NPupBar.refNPupBar.setState({messages: Object.assign({}, this.messages)});